refactor(structure): normalize folder naming convention
- Rename Layout -> layout
- Rename Section -> sections
- Update import paths to match new folder structure

This change standardizes the folder naming convention to lowercase
for better consistency with React/Next.js best practices.
