+++ src/components/common/Magnetic.tsx
