+++ src/components/common/SliderText.tsx
