+++ src/components/common/TechSkills.tsx
