+++ src/components/layout/Curve.tsx
