+++ src/components/sections/Projects.tsx
@@ -3,12 +3,12 @@
 import { useState, useEffect, useRef } from 'react';
 import { useInView } from 'react-intersection-observer';
 import { motion, useScroll } from 'framer-motion';
-import Project from '@/common/ProjectItem';
+import Project from '@/components/common/ProjectItem';
 import { projects } from '@/types/projects';
 import { AnimatePresence } from 'framer-motion';
 import gsap from 'gsap';
 import Image from 'next/image';
-import Rounded from '@/common/RoundedButton';
+import Rounded from '@/components/common/RoundedButton';
 
 const scaleAnimation = {
   initial: {scale: 0, x:"-50%", y:"-50%"},
