+++ src/components/sections/SlideImage.tsx
